determine erroriness inside evaluate
The function returns both the int status_code and a bool for whether to
consider the status an error.

Also, refactors the guts of the function to a switch statement. I think
this reads a little clearer while we've got special cases for multiple
enums. When the deprecation goes away, this can probably revert back to
simpler if statements.