Convert OTLP span status.code type to integer earlier in the transform (#78)
## Which problem is this PR solving?

**Sign/Symptom:** 

A customer reported that `status_code` values had stopped appearing abruptly in one of their datasets. They had changed the type of the `status_code` column in the dataset schema from string to integer. _This is a totally reasonable thing to do because an OpenTelemetry span status code is an integer._

**Origin:**

Prior to this change, Husky sets `status_code` to a `trace.Status_StatusCode` type from the trace.pb.go generated from OTLP proto definitions. For Reasons™, that type _would_ convert to a string during our data ingest, but would _not_ convert to an integer (despite the type's association with actual `int32` constants).

## Short description of the changes

We'll now do the type conversion from OTLP proto enum shenanigans to an `int` here in husky's transform of an OTLP span into an event. The events we produce from OTLP spans have no knowledge of or interest in the OTLP types generated from enums in the proto definitions.

## Longer description of the changes

* appease linter with use of deprecated code

* rename and add tests 

* evaluate status once, then reuse

* convert status_code to int & determine erroriness inside evaluate

The function returns both the int status_code and a bool for whether to
consider the status an error.

Also, refactors the guts of the function to a switch statement. I think
this reads a little clearer while we've got special cases for multiple
enums. When the deprecation goes away, this can probably revert back to
simpler if statements.

Co-authored-by: Mike Goldsmith <mikegoldsmith@honeycomb.io>