update trace ingest tests to support JSON
Consolidates the different Legacy/NonLegacy test functions into
test cases within a single function. The majority of test assertions are
the same except for what headers are used for the request and what
dataset is targeted for the incoming telemetry.

Test cases updated to use the common_testhelpers to loop over supported
content types and encodings.