[feat] allow OTLP/http-json for metrics (#121)
Enabling support for all the OTLP content types. The ingest for
metrics is still handled elsewhere at the moment, but that process
uses husky's header validation before proceeding, so allowing
JSON needs to happen here first.

* Refactored how supported content types are checked.
  A list of supported types is declared once and public function is
  provided to check for the presence of a given type in the supported
  list.

* Bonus: included a correction to the HTTP error we return when a
  content type is not supported: 415 Unsupported Media Type instead
  of 501 Not Implemented. If we don't support a content type, that's
  not a server side error.

* Refactored ingestion tests into loop-driven test cases of Classic/E&S,
  content types, and content encodings.

* update trace ingest tests to support JSON

Consolidates the different Legacy/NonLegacy test functions into
test cases within a single function. The majority of test assertions are
the same except for what headers are used for the request and what
dataset is targeted for the incoming telemetry.

Test cases updated to use the common_testhelpers to loop over supported
content types and encodings.

* Reorg and consolidate header validation as series of if-guards. 
  Currently all OTLP signal requests have the same validation rules.
  This consolidates those rules to a single private function. The public
  API retains the separate, signal-specific validation functions to allow for
  potential signal-specific validations returning in the future and to
  reduce code churn between now and then.