OTLP size fix (#146)
## Which problem is this PR solving?

- Part 2 of #145 
- This imposes a field size limit on values inbound from OTLP by cutting
them off at a size limit in the marshaling process. This will cause them
to fail to unmarshal as JSON, but since they're being sent as strings
anyway, it doesn't seem like a very big deal.
- The size limit is 64K bytes, which is also imposed by Shepherd, but
there's no point in doing the work if we're going to trim it anyway.
- Closes #147 

## Short description of the changes

- A new limitedWriter type wraps strings.Builder for efficiency, and
will only write up to the size limit. After reaching it, it lies to the
caller and pretends it writes, but it doesn't.
- Because we had to change to a json.Encoder to use this writer, we also
have to cope with the fact that json.Encoder terminates its results with
a newline, so some tests had to be changed to accommodate it.
- I tried to use a different, supposedly faster, json encoder but it
broke several tests so I immediately abandoned it.
- I added some telemetry to see when this is being invoked -- it adds
the fields `meta.truncated_bytes` and `meta.truncated_field` to the span
if a field is truncated. This makes it visible to customers. We might
also want to consider adding something visible in Dogfood if this
happens but that's not a Husky change.
- I also took the recommendation in #147 to use a faster JSON encoder
since we're already using it in Shepherd.

Fixes #147.