fix: properly parse parent id for logs (#230)
## Which problem is this PR solving?

- In https://github.com/honeycombio/husky/pull/179 we added some code to
deal with the fact that OTLP/JSON processing can't differentiate between
arrays of bytes that are encoded as hex and those that are encoded as
base64. When JS sends us browser-side telemetry, it encodes IDs as hex,
but when we receive it it's decoded as base64. This PR also applies that
to logs.

## Short description of the changes

- Adjust log's parentID so that it properly parses
- Add a new test case
- Move some code to common since it now applies across signals
- Move some tests to common
- Clarify some comments