feat: support JSON OTLP in new fast path (#308)
## Which problem is this PR solving?
JSON is still a thing even in 2025 - some SDKs ONLY support sending the
JSON flavor of OTLP. We'll need to handle this efficiently.

## Short description of the changes
This extends the work in #307 with a JSON version mirroring the proto
version. It's not great to have, in effect, two copies of the same
business logic, but I think the language gymnastics we'd have to do to
consolidate it between both implementations in an efficient way would
make the code truly impenetrable. We'll have to have a lot of faith in
our testing - `TestUnmarshalTraceRequestDirect_Complete` now includes
JSON.

I'm using fastjson because jsoniter doesn't actually support
zero-allocation object iteration. Unfortunately both libraries appear to
be abandonware, with no modern alternative on the horizon.

This is much slower than protobuf, but that's to be expected, and at
least it's in the same ballpark.
```
BenchmarkUnmarshalTraceRequestDirectMsgp/Protobuf-12         	  156210	      6451 ns/op	    3377 B/op	       5 allocs/op
BenchmarkUnmarshalTraceRequestDirectMsgp/JSON-12             	   71346	     16525 ns/op	    3430 B/op	       7 allocs/op
```