feat: export UnmarshalTraceRequestDirectMsgp (#309)
## Which problem is this PR solving?
TranslateTraceRequestFromReaderSizedWithMsgp is designed for HTTP
endpoints, where you have a reader and a content type. In GRPC-land,
you're handed an already-read []byte which is assumed to be protobuf, so
we can skip a level of indirection and buffering if we go directly to
the unmarshal function for GRPC requests.

## Short description of the changes
Just renames the function.