feat: add prefix used by Collector telemetry scopes to recognized list (#314)
## Which problem is this PR solving?

We would like to be able to assess the provenance of telemetry sent to
Honeycomb. We maintain an allowlist of strings that identify the open
source OpenTelemetry instrumentation libraries. The allowlist lets us
copy instrumentation_scope.name values from incoming telemetry, but only
for the OTel instrumentation, not custom names given to custom tracers,
meters, or loggers.

## Short description of the changes

- relocate test: this logic is in common, not traces
- renamed test property to something other than the name of the function
under test
- add the prefix that identifies telemetry coming from the OpenTelemetry
Collector's instrumentation

I opted to stop the prefix at "-collector" and not "-collector-contrib"
because the shorter is:

1. sufficient to identify the open source OTel stuff
2. who knows, maybe things other than "-contrib" get appended later, but
it'll still be OTel's